Create entry types file to hold TabLocation.
