+++ src/shared/components/error-pages/ObsoleteEntryPage.tsx
@@ -9,8 +9,7 @@ import {
   getEntryPath,
 } from '../../../app/config/urls';
 import { Namespace } from '../../types/namespaces';
-
-import { TabLocation } from '../../../uniprotkb/components/entry/Entry';
+import { TabLocation } from '../../../uniprotkb/types/entry';
 
 const ObsoleteMessageHeading = ({ release }: { release?: string }) => (
   <h4>This entry is obsolete{release && ` since release ${release}`}</h4>
