+++ src/uniprotkb/components/entry/entry.structured.ts
@@ -3,7 +3,7 @@ import { WithContext } from 'schema-dts';
 import { getEntryPath } from '../../../app/config/urls';
 
 import { Namespace } from '../../../shared/types/namespaces';
-import { TabLocation } from './Entry';
+import { TabLocation } from '../../types/entry';
 
 import { UniProtkbAPIModel } from '../../adapters/uniProtkbConverter';
