+++ src/uniprotkb/types/entry.ts
@@ -1,0 +1,9 @@
+export enum TabLocation {
+  Entry = 'entry',
+  VariantViewer = 'variant-viewer',
+  FeatureViewer = 'feature-viewer',
+  GenomicCoordinates = 'genomic-coordinates',
+  Publications = 'publications',
+  ExternalLinks = 'external-links',
+  History = 'history',
+}
